lhx_text2sql/requirements.txt
pydantic==1.10.13
jsonschema==4.19.2
sqlglot==23.8.1
pymysql==1.1.0
pytest==7.4.3
PyQt5==5.15.10
python-dotenv==1.0.1
